+++ serialization/src/commonMain/kotlin/nl/adaptivity/xmlutil/serialization/XML.kt
@@ -1229,6 +1229,15 @@ internal fun XmlChildrenName.toQName(parentNamespace: Namespace?): QName = when 
     else -> QName(namespace, value, prefix)
 }
 
+internal fun XmlKeyName.toQName(parentNamespace: Namespace?): QName = when {
+    namespace == UNSET_ANNOTATION_VALUE -> parentNamespace?.let { QName(it.namespaceURI, value) } ?: QName(value)
+    prefix == UNSET_ANNOTATION_VALUE -> {
+        val p = parentNamespace?.let { ns -> ns.prefix.takeIf { ns.namespaceURI == namespace }}
+        QName(namespace, value, p ?: "")
+    }
+    else -> QName(namespace, value, prefix)
+}
+
 internal inline fun <reified T> Iterable<*>.firstOrNull(): T? {
     for (e in this) {
         if (e is T) return e